fix: 优化页面展示
