+++ src/app/blog/@main/[cate_id]/[post_id]/page.tsx
@@ -27,7 +27,7 @@ export default async function Post({ params }: Props) {
   const currentArticle = await getArticle(post_id, cate_id)
 
   return (
-    <div className="flex min-h-screen flex-col items-center justify-between p-24">
+    <div className="flex min-h-screen flex-col items-center justify-between px-24">
       <Editor
         readOnly
         articleId={post_id}
