+++ src/app/dashboard/@sidebar/page.tsx
@@ -1,15 +1,5 @@
 'use client'
-import { Category } from '@/server/db/schema'
-import { useSidebarStore } from '@/hooks'
-import { MenuType } from '@/types/Common'
-import { Sidebar } from '@/types/sidebar'
-import { trpc } from '@/utils/trpc-client'
-import { ScrollShadow } from '@nextui-org/react'
-import { createId } from '@paralleldrive/cuid2'
-import { useDebounceFn } from 'ahooks'
-import { useParams, usePathname } from 'next/navigation'
-import { useEffect, useState } from 'react'
-import { shallow } from 'zustand/shallow'
+
 import { Collapse } from '@/components/collapse/collapse'
 import { AddArticle } from '@/components/sidebar/add-article-action'
 import { ArticleTitle } from '@/components/sidebar/article-title'
@@ -18,21 +8,24 @@ import { PopoverInput } from '@/components/sidebar/popover-input'
 import { SidebarItem } from '@/components/sidebar/sidebar-item'
 import { MenuTitle } from '@/components/sidebar/sidebar-menu'
 import { SidebarTop } from '@/components/sidebar/sidebar-top'
-
-type MenuParam = Pick<Category, 'id' | 'name'>
-
-type DiffResult = {
-  add: MenuParam[]
-  modi: MenuParam[]
-  del: MenuParam[]
-}
+import { useSidebarStore } from '@/hooks'
+import { Category } from '@/server/db/schema'
+import { MenuType } from '@/types/Common'
+import { diffChildren } from '@/utils/sidebar'
+import { trpc } from '@/utils/trpc-client'
+import { ScrollShadow } from '@nextui-org/react'
+import { createId } from '@paralleldrive/cuid2'
+import { useDebounceFn } from 'ahooks'
+import { useResponsive } from 'ahooks/lib/useResponsive'
+import { motion } from 'framer-motion'
+import { useParams, usePathname } from 'next/navigation'
+import { useEffect, useState } from 'react'
+import { shallow } from 'zustand/shallow'
 
 export default function SideBarPage() {
   const pathName = usePathname()
   const { slugs } = useParams<{ slugs: string[] }>()
-
   const [cateId, articleId] = slugs || []
-
   const [activeId, setActiveId] = useState<string>('')
   const [hoverId, setHoverId] = useState<string>('')
   const [showDropdownId, setShowDropdownId] = useState<string>('')
@@ -62,6 +55,7 @@ export default function SideBarPage() {
   const { mutate: deleteCategoryById } = trpc.deleteCategoryById.useMutation()
 
   const { run: updateCategoryDebounce } = useDebounceFn(updateCategory)
+  const { md } = useResponsive()
 
   useEffect(() => {
     if (categoriesFromServer) {
@@ -116,43 +110,6 @@ export default function SideBarPage() {
     )
   }, [])
 
-  const diffChildren = (curChildren: Sidebar[], preChildren: Sidebar[]) => {
-    const diff: DiffResult = {
-      add: [],
-      del: [],
-      modi: [],
-    }
-    const preArray = Array.from(preChildren)
-    curChildren.forEach((cur) => {
-      const findPre = preArray.find((pre) => {
-        return cur.id === pre.id
-      })
-      const curOnPreIndex = preArray.findIndex((pre) => {
-        return cur.id === pre.id
-      })
-      if (curOnPreIndex !== -1) {
-        preArray.splice(curOnPreIndex, 1)
-      }
-      if (!findPre) {
-        console.log('cur children add')
-        diff.add.push({
-          name: cur.name as string,
-          id: cur.id,
-        })
-      } else {
-        if (findPre.name !== cur.name) {
-          console.log('cur children update')
-          diff.modi.push({
-            name: cur.name as string,
-            id: cur.id,
-          })
-        }
-      }
-    })
-    diff.del = preArray as MenuParam[]
-    return diff
-  }
-
   const onAddHandle = () => {
     insertMenu({
       id: createId(),
@@ -183,91 +140,94 @@ export default function SideBarPage() {
   }
 
   return (
-    <ScrollShadow
+    <motion.div
+      animate={{ width: !md ? '0' : '240px' }}
       className="sticky left-0 top-0 box-border  
     flex h-full w-72 flex-col space-y-2
      border-r-1 border-default-100 px-4"
     >
-      <SidebarTop />
-      {sidebars.map((sidebar) => {
-        return (
-          <div key={sidebar.id}>
-            <MenuTitle onAdd={onAddHandle} id={sidebar.id}>
-              {sidebar.name}
-            </MenuTitle>
-            {Array.isArray(sidebar.children) &&
-              sidebar.children.map((subSidebar) => {
-                if (Array.isArray(subSidebar.children)) {
-                  return (
-                    <Collapse
-                      id={subSidebar.id}
-                      items={subSidebar.children?.map((child) => {
-                        return (
-                          <ArticleTitle
-                            onSidebarDel={() => {
-                              deleteArticleTitleById(subSidebar.id, child.id)
-                              delArticleById({
-                                id: child.id,
-                              })
-                            }}
-                            isActive={child.id === articleId}
-                            key={child.id}
-                            href={child.href || ''}
-                          >
-                            {child.name || 'no title'}
-                          </ArticleTitle>
-                        )
-                      })}
-                      isActived={cateId === subSidebar.id}
-                      startContent={subSidebar.icon || ''}
-                      key={subSidebar.id}
-                      onCollapseChange={onCollapseChange}
-                      onHover={setHoverId}
-                      title={
-                        <div className="flex w-full items-center justify-between">
-                          <PopoverInput
-                            onPopoverInputChange={editMenu}
-                            onClose={() => setActiveId('')}
-                            id={subSidebar.id}
-                            key={subSidebar.id + 'PopoverInput'}
-                            name={subSidebar.name as string}
-                            activeId={activeId}
-                          ></PopoverInput>
+      <ScrollShadow>
+        <SidebarTop />
+        {sidebars.map((sidebar) => {
+          return (
+            <div key={sidebar.id}>
+              <MenuTitle onAdd={onAddHandle} id={sidebar.id}>
+                {sidebar.name}
+              </MenuTitle>
+              {Array.isArray(sidebar.children) &&
+                sidebar.children.map((subSidebar) => {
+                  if (Array.isArray(subSidebar.children)) {
+                    return (
+                      <Collapse
+                        id={subSidebar.id}
+                        items={subSidebar.children?.map((child) => {
+                          return (
+                            <ArticleTitle
+                              onSidebarDel={() => {
+                                deleteArticleTitleById(subSidebar.id, child.id)
+                                delArticleById({
+                                  id: child.id,
+                                })
+                              }}
+                              isActive={child.id === articleId}
+                              key={child.id}
+                              href={child.href || ''}
+                            >
+                              {child.name || 'no title'}
+                            </ArticleTitle>
+                          )
+                        })}
+                        isActived={cateId === subSidebar.id}
+                        startContent={subSidebar.icon || ''}
+                        key={subSidebar.id}
+                        onCollapseChange={onCollapseChange}
+                        onHover={setHoverId}
+                        title={
+                          <div className="flex w-full items-center justify-between">
+                            <PopoverInput
+                              onPopoverInputChange={editMenu}
+                              onClose={() => setActiveId('')}
+                              id={subSidebar.id}
+                              key={subSidebar.id + 'PopoverInput'}
+                              name={subSidebar.name as string}
+                              activeId={activeId}
+                            ></PopoverInput>
 
-                          {hoverId === subSidebar.id ||
-                          showDropdownId === subSidebar.id ? (
-                            <div>
-                              <DropDownMenu
-                                key={subSidebar.id + 'DropDownMenu'}
-                                onOpenChange={setShowDropdownId}
-                                id={subSidebar.id}
-                                onAction={onDropdownHandle}
-                              ></DropDownMenu>
-                              <AddArticle
-                                categoryId={subSidebar.id}
-                                onAddArticle={onAddArticle}
-                              />
-                            </div>
-                          ) : null}
-                        </div>
-                      }
-                    ></Collapse>
+                            {hoverId === subSidebar.id ||
+                            showDropdownId === subSidebar.id ? (
+                              <div>
+                                <DropDownMenu
+                                  key={subSidebar.id + 'DropDownMenu'}
+                                  onOpenChange={setShowDropdownId}
+                                  id={subSidebar.id}
+                                  onAction={onDropdownHandle}
+                                ></DropDownMenu>
+                                <AddArticle
+                                  categoryId={subSidebar.id}
+                                  onAddArticle={onAddArticle}
+                                />
+                              </div>
+                            ) : null}
+                          </div>
+                        }
+                      ></Collapse>
+                    )
+                  }
+                  return (
+                    <SidebarItem
+                      isActive={pathName === subSidebar.href}
+                      key={subSidebar.id}
+                      href={subSidebar.href || ''}
+                      icon={subSidebar.icon}
+                    >
+                      {subSidebar.name}
+                    </SidebarItem>
                   )
-                }
-                return (
-                  <SidebarItem
-                    isActive={pathName === subSidebar.href}
-                    key={subSidebar.id}
-                    href={subSidebar.href || ''}
-                    icon={subSidebar.icon}
-                  >
-                    {subSidebar.name}
-                  </SidebarItem>
-                )
-              })}
-          </div>
-        )
-      })}
-    </ScrollShadow>
+                })}
+            </div>
+          )
+        })}
+      </ScrollShadow>
+    </motion.div>
   )
 }
